refactor: dividing the files into packages for clarity
